+++ cypress/e2e/data-import/e2e/field-protection-overrides.cy.js
@@ -1,25 +1,25 @@
-import getRandomPostfix from '../../support/utils/stringTools';
-import TestTypes from '../../support/dictionary/testTypes';
-import DevTeams from '../../support/dictionary/devTeams';
-import SettingsMenu from '../../support/fragments/settingsMenu';
-import NewFieldMappingProfile from '../../support/fragments/data_import/mapping_profiles/newFieldMappingProfile';
-import FieldMappingProfiles from '../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
-import MarcFieldProtection from '../../support/fragments/settings/dataImport/marcFieldProtection';
-import FieldMappingProfileView from '../../support/fragments/data_import/mapping_profiles/fieldMappingProfileView';
-import ActionProfiles from '../../support/fragments/data_import/action_profiles/actionProfiles';
-import NewActionProfile from '../../support/fragments/data_import/action_profiles/newActionProfile';
-import MatchProfiles from '../../support/fragments/data_import/match_profiles/matchProfiles';
-import NewJobProfile from '../../support/fragments/data_import/job_profiles/newJobProfile';
-import JobProfiles from '../../support/fragments/data_import/job_profiles/jobProfiles';
-import TopMenu from '../../support/fragments/topMenu';
-import DataImport from '../../support/fragments/data_import/dataImport';
-import FileDetails from '../../support/fragments/data_import/logs/fileDetails';
-import Logs from '../../support/fragments/data_import/logs/logs';
-import FileManager from '../../support/utils/fileManager';
-import InstanceRecordView from '../../support/fragments/inventory/instanceRecordView';
-import InventoryInstance from '../../support/fragments/inventory/inventoryInstance';
-import InventoryViewSource from '../../support/fragments/inventory/inventoryViewSource';
-import InventorySearchAndFilter from '../../support/fragments/inventory/inventorySearchAndFilter';
+import getRandomPostfix from '../../../support/utils/stringTools';
+import TestTypes from '../../../support/dictionary/testTypes';
+import DevTeams from '../../../support/dictionary/devTeams';
+import SettingsMenu from '../../../support/fragments/settingsMenu';
+import NewFieldMappingProfile from '../../../support/fragments/data_import/mapping_profiles/newFieldMappingProfile';
+import FieldMappingProfiles from '../../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
+import MarcFieldProtection from '../../../support/fragments/settings/dataImport/marcFieldProtection';
+import FieldMappingProfileView from '../../../support/fragments/data_import/mapping_profiles/fieldMappingProfileView';
+import ActionProfiles from '../../../support/fragments/data_import/action_profiles/actionProfiles';
+import NewActionProfile from '../../../support/fragments/data_import/action_profiles/newActionProfile';
+import MatchProfiles from '../../../support/fragments/data_import/match_profiles/matchProfiles';
+import NewJobProfile from '../../../support/fragments/data_import/job_profiles/newJobProfile';
+import JobProfiles from '../../../support/fragments/data_import/job_profiles/jobProfiles';
+import TopMenu from '../../../support/fragments/topMenu';
+import DataImport from '../../../support/fragments/data_import/dataImport';
+import FileDetails from '../../../support/fragments/data_import/logs/fileDetails';
+import Logs from '../../../support/fragments/data_import/logs/logs';
+import FileManager from '../../../support/utils/fileManager';
+import InstanceRecordView from '../../../support/fragments/inventory/instanceRecordView';
+import InventoryInstance from '../../../support/fragments/inventory/inventoryInstance';
+import InventoryViewSource from '../../../support/fragments/inventory/inventoryViewSource';
+import InventorySearchAndFilter from '../../../support/fragments/inventory/inventorySearchAndFilter';
 
 describe('ui-data-import: Check that field protection overrides work properly during data import', () => {
   let firstFieldId = null;
