+++ cypress/e2e/data-import/e2e/field-protection-settings.cy.js
@@ -1,25 +1,25 @@
-import getRandomPostfix from '../../support/utils/stringTools';
-import TestTypes from '../../support/dictionary/testTypes';
-import DevTeams from '../../support/dictionary/devTeams';
-import NewFieldMappingProfile from '../../support/fragments/data_import/mapping_profiles/newFieldMappingProfile';
-import FieldMappingProfiles from '../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
-import SettingsMenu from '../../support/fragments/settingsMenu';
-import NewActionProfile from '../../support/fragments/data_import/action_profiles/newActionProfile';
-import ActionProfiles from '../../support/fragments/data_import/action_profiles/actionProfiles';
-import NewJobProfile from '../../support/fragments/data_import/job_profiles/newJobProfile';
-import JobProfiles from '../../support/fragments/data_import/job_profiles/jobProfiles';
-import TopMenu from '../../support/fragments/topMenu';
-import DataImport from '../../support/fragments/data_import/dataImport';
-import Logs from '../../support/fragments/data_import/logs/logs';
-import FileDetails from '../../support/fragments/data_import/logs/fileDetails';
-import InstanceRecordView from '../../support/fragments/inventory/instanceRecordView';
-import MarcFieldProtection from '../../support/fragments/settings/dataImport/marcFieldProtection';
-import InventorySearchAndFilter from '../../support/fragments/inventory/inventorySearchAndFilter';
-import MatchProfiles from '../../support/fragments/data_import/match_profiles/matchProfiles';
-import NewMatchProfile from '../../support/fragments/data_import/match_profiles/newMatchProfile';
-import InventoryViewSource from '../../support/fragments/inventory/inventoryViewSource';
-import FileManager from '../../support/utils/fileManager';
-import InventoryInstance from '../../support/fragments/inventory/inventoryInstance';
+import getRandomPostfix from '../../../support/utils/stringTools';
+import TestTypes from '../../../support/dictionary/testTypes';
+import DevTeams from '../../../support/dictionary/devTeams';
+import NewFieldMappingProfile from '../../../support/fragments/data_import/mapping_profiles/newFieldMappingProfile';
+import FieldMappingProfiles from '../../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
+import SettingsMenu from '../../../support/fragments/settingsMenu';
+import NewActionProfile from '../../../support/fragments/data_import/action_profiles/newActionProfile';
+import ActionProfiles from '../../../support/fragments/data_import/action_profiles/actionProfiles';
+import NewJobProfile from '../../../support/fragments/data_import/job_profiles/newJobProfile';
+import JobProfiles from '../../../support/fragments/data_import/job_profiles/jobProfiles';
+import TopMenu from '../../../support/fragments/topMenu';
+import DataImport from '../../../support/fragments/data_import/dataImport';
+import Logs from '../../../support/fragments/data_import/logs/logs';
+import FileDetails from '../../../support/fragments/data_import/logs/fileDetails';
+import InstanceRecordView from '../../../support/fragments/inventory/instanceRecordView';
+import MarcFieldProtection from '../../../support/fragments/settings/dataImport/marcFieldProtection';
+import InventorySearchAndFilter from '../../../support/fragments/inventory/inventorySearchAndFilter';
+import MatchProfiles from '../../../support/fragments/data_import/match_profiles/matchProfiles';
+import NewMatchProfile from '../../../support/fragments/data_import/match_profiles/newMatchProfile';
+import InventoryViewSource from '../../../support/fragments/inventory/inventoryViewSource';
+import FileManager from '../../../support/utils/fileManager';
+import InventoryInstance from '../../../support/fragments/inventory/inventoryInstance';
 
 describe('ui-data-import: Check that field protection settings work properly during data import', () => {
   const marcFieldProtected = ['507', '920'];
