+++ cypress/e2e/data-import/e2e/match-on-holdings-856-u.cy.js
@@ -1,20 +1,20 @@
-import getRandomPostfix from '../../support/utils/stringTools';
-import DevTeams from '../../support/dictionary/devTeams';
-import TestTypes from '../../support/dictionary/testTypes';
-import ActionProfiles from '../../support/fragments/data_import/action_profiles/actionProfiles';
-import DataImport from '../../support/fragments/data_import/dataImport';
-import JobProfiles from '../../support/fragments/data_import/job_profiles/jobProfiles';
-import NewJobProfile from '../../support/fragments/data_import/job_profiles/newJobProfile';
-import FieldMappingProfiles from '../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
-import NewFieldMappingProfile from '../../support/fragments/data_import/mapping_profiles/newFieldMappingProfile';
-import MatchProfiles from '../../support/fragments/data_import/match_profiles/matchProfiles';
-import NewMatchProfile from '../../support/fragments/data_import/match_profiles/newMatchProfile';
-import HoldingsRecordView from '../../support/fragments/inventory/holdingsRecordView';
-import InventoryInstance from '../../support/fragments/inventory/inventoryInstance';
-import SettingsMenu from '../../support/fragments/settingsMenu';
-import TopMenu from '../../support/fragments/topMenu';
-import Logs from '../../support/fragments/data_import/logs/logs';
-import InventorySearchAndFilter from '../../support/fragments/inventory/inventorySearchAndFilter';
+import getRandomPostfix from '../../../support/utils/stringTools';
+import DevTeams from '../../../support/dictionary/devTeams';
+import TestTypes from '../../../support/dictionary/testTypes';
+import ActionProfiles from '../../../support/fragments/data_import/action_profiles/actionProfiles';
+import DataImport from '../../../support/fragments/data_import/dataImport';
+import JobProfiles from '../../../support/fragments/data_import/job_profiles/jobProfiles';
+import NewJobProfile from '../../../support/fragments/data_import/job_profiles/newJobProfile';
+import FieldMappingProfiles from '../../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
+import NewFieldMappingProfile from '../../../support/fragments/data_import/mapping_profiles/newFieldMappingProfile';
+import MatchProfiles from '../../../support/fragments/data_import/match_profiles/matchProfiles';
+import NewMatchProfile from '../../../support/fragments/data_import/match_profiles/newMatchProfile';
+import HoldingsRecordView from '../../../support/fragments/inventory/holdingsRecordView';
+import InventoryInstance from '../../../support/fragments/inventory/inventoryInstance';
+import SettingsMenu from '../../../support/fragments/settingsMenu';
+import TopMenu from '../../../support/fragments/topMenu';
+import Logs from '../../../support/fragments/data_import/logs/logs';
+import InventorySearchAndFilter from '../../../support/fragments/inventory/inventorySearchAndFilter';
 
 describe('ui-data-import: Match on Holdings 856 $u', () => {
   const matchProfileName = `autotestMatchProf${getRandomPostfix()}`;
