+++ cypress/e2e/data-import/e2e/match-on-instance-identifier-scen2.cy.js
@@ -1,32 +1,34 @@
-import TestTypes from '../../support/dictionary/testTypes';
-import DevTeams from '../../support/dictionary/devTeams';
-import TopMenu from '../../support/fragments/topMenu';
-import getRandomPostfix from '../../support/utils/stringTools';
-import DataImport from '../../support/fragments/data_import/dataImport';
-import JobProfiles from '../../support/fragments/data_import/job_profiles/jobProfiles';
-import Logs from '../../support/fragments/data_import/logs/logs';
-import InventoryInstance from '../../support/fragments/inventory/inventoryInstance';
-import SettingsMenu from '../../support/fragments/settingsMenu';
-import MatchProfiles from '../../support/fragments/data_import/match_profiles/matchProfiles';
-import FieldMappingProfiles from '../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
-import NewActionProfile from '../../support/fragments/data_import/action_profiles/newActionProfile';
-import ActionProfiles from '../../support/fragments/data_import/action_profiles/actionProfiles';
-import NewJobProfile from '../../support/fragments/data_import/job_profiles/newJobProfile';
-import InstanceRecordView from '../../support/fragments/inventory/instanceRecordView';
-import InventorySearchAndFilter from '../../support/fragments/inventory/inventorySearchAndFilter';
-import permissions from '../../support/dictionary/permissions';
-import Users from '../../support/fragments/users/users';
+import TestTypes from '../../../support/dictionary/testTypes';
+import DevTeams from '../../../support/dictionary/devTeams';
+import TopMenu from '../../../support/fragments/topMenu';
+import getRandomPostfix from '../../../support/utils/stringTools';
+import DataImport from '../../../support/fragments/data_import/dataImport';
+import JobProfiles from '../../../support/fragments/data_import/job_profiles/jobProfiles';
+import Logs from '../../../support/fragments/data_import/logs/logs';
+import InventoryInstance from '../../../support/fragments/inventory/inventoryInstance';
+import SettingsMenu from '../../../support/fragments/settingsMenu';
+import MatchProfiles from '../../../support/fragments/data_import/match_profiles/matchProfiles';
+import FieldMappingProfiles from '../../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
+import NewActionProfile from '../../../support/fragments/data_import/action_profiles/newActionProfile';
+import ActionProfiles from '../../../support/fragments/data_import/action_profiles/actionProfiles';
+import NewJobProfile from '../../../support/fragments/data_import/job_profiles/newJobProfile';
+import InstanceRecordView from '../../../support/fragments/inventory/instanceRecordView';
+import InventorySearchAndFilter from '../../../support/fragments/inventory/inventorySearchAndFilter';
+import permissions from '../../../support/dictionary/permissions';
+import Users from '../../../support/fragments/users/users';
 
-describe('ui-data-import: Match on Instance identifier match meets both the Identifier type and Data requirements (Scenario 1)', () => {
+describe('ui-data-import: Match on Instance identifier match meets both the Identifier type and Data requirements (Scenario 2)', () => {
   let userId;
-  const fileNameForCreateInstance = `C347828autotestFile.${getRandomPostfix()}.mrc`;
-  const fileNameForUpdateInstance = `C347828autotestFile.${getRandomPostfix()}.mrc`;
+  const filePathForCreateInstance = 'marcFileForMatchOnIdentifierForCreate.mrc';
+  const filePathForUpdateInstance = 'marcFileForMatchOnIdentifierForUpdate_2.mrc';
+  const fileNameForCreateInstance = `C358137autotestFile.${getRandomPostfix()}.mrc`;
+  const fileNameForUpdateInstance = `C358137autotestFile.${getRandomPostfix()}.mrc`;
   const jobProfileToRun = 'Default - Create instance and SRS MARC Bib';
   const matchProfileName = `autotestMatchProf${getRandomPostfix()}`;
   const mappingProfileName = `autotestMappingProf${getRandomPostfix()}`;
   const actionProfileName = `autotestActionProf${getRandomPostfix()}`;
-  const jobProfileName = `autotestJobProf${getRandomPostfix()}`;
-  const instanceGeneralNote = 'IDENTIFIER UPDATE 1';
+  const jobProfileName = `autotestActionProf${getRandomPostfix()}`;
+  const instanceGeneralNote = 'IDENTIFIER UPDATE 2';
   const resourceIdentifiers = [
     { type: 'UPC', value: 'ORD32671387-4' },
     { type: 'OCLC', value: '(OCoLC)84714376518561876438' },
@@ -38,34 +40,34 @@ describe('ui-data-import: Match on Instance identifier match meets both the Iden
     incomingRecordFields: {
       field: '024',
       in1: '1',
-      in2: '*',
-      subfield: 'a'
+      subfield: 'z'
     },
     matchCriterion: 'Exactly matches',
     existingRecordType: 'INSTANCE',
-    instanceOption: 'Identifier: UPC',
+    instanceOption: 'Identifier: Invalid UPC',
   };
   const mappingProfile = {
     name: mappingProfileName,
     incomingRecordType: 'MARC Bibliographic',
     folioRecordType: 'Instance',
-    discoverySuppress: 'Mark for all affected records',
-    catalogedDate: '"2021-12-01"',
-    catalogedDateUI: '2021-12-01',
-    instanceStatus: '"Batch Loaded"',
-    instanceStatusUI: 'Batch Loaded',
+    staffSuppress: 'Mark for all affected records',
+    catalogedDate: '"2021-12-02"',
+    catalogedDateUI: '2021-12-02',
+    instanceStatus: '"Cataloged"',
+    instanceStatusUI: 'Cataloged',
   };
   const actionProfile = {
-    name: actionProfileName,
     typeValue : NewActionProfile.folioRecordTypeValue.instance,
+    name: actionProfileName,
     action: 'Update (all record types except Orders, Invoices, or MARC Holdings)'
   };
   const jobProfile = {
+    ...NewJobProfile.defaultJobProfile,
     profileName: jobProfileName,
     acceptedType: NewJobProfile.acceptedDataType.marc
   };
 
-  before(() => {
+  beforeEach(() => {
     cy.createTempUser([
       permissions.moduleDataImportEnabled.gui,
       permissions.dataImportDeleteLogs.gui,
@@ -82,27 +84,22 @@ describe('ui-data-import: Match on Instance identifier match meets both the Iden
           path: TopMenu.dataImportPath,
           waiter: DataImport.waitLoading
         });
+      });
 
-        InventorySearchAndFilter.getInstancesByIdentifierViaApi(resourceIdentifiers[0].value)
-          .then(instances => {
-            instances.forEach(({ id }) => {
-              InventoryInstance.deleteInstanceViaApi(id);
-            });
-          });
+    InventorySearchAndFilter.getInstancesByIdentifierViaApi(resourceIdentifiers[0].value)
+      .then(instances => {
+        instances.forEach(({ id }) => {
+          InventoryInstance.deleteInstanceViaApi(id);
+        });
       });
   });
 
   after(() => {
     Users.deleteViaApi(userId);
-    // delete profiles
-    JobProfiles.deleteJobProfile(jobProfileName);
-    MatchProfiles.deleteMatchProfile(matchProfileName);
-    ActionProfiles.deleteActionProfile(actionProfileName);
-    FieldMappingProfiles.deleteFieldMappingProfile(mappingProfileName);
   });
 
-  it('C347828 Match on Instance identifier match meets both the Identifier type and Data requirements (folijet)', { tags: [TestTypes.criticalPath, DevTeams.folijet] }, () => {
-    DataImport.uploadFile('marcFileForMatchOnIdentifierForCreate.mrc', fileNameForCreateInstance);
+  it('C347829 Match on Instance identifier match meets both the Identifier type and Data requirements (folijet)', { tags: [TestTypes.criticalPath, DevTeams.folijet] }, () => {
+    DataImport.uploadFile(filePathForCreateInstance, fileNameForCreateInstance);
     JobProfiles.searchJobProfileForImport(jobProfileToRun);
     JobProfiles.runImportFile(fileNameForCreateInstance);
     Logs.checkStatusOfJobProfile('Completed');
@@ -117,11 +114,9 @@ describe('ui-data-import: Match on Instance identifier match meets both the Iden
 
     cy.visit(SettingsMenu.matchProfilePath);
     MatchProfiles.createMatchProfile(matchProfile);
-    MatchProfiles.checkMatchProfilePresented(matchProfileName);
 
     cy.visit(SettingsMenu.mappingProfilePath);
     FieldMappingProfiles.createMappingProfileForMatchOnInstanceIdentifier(mappingProfile);
-    FieldMappingProfiles.checkMappingProfilePresented(mappingProfileName);
 
     cy.visit(SettingsMenu.actionProfilePath);
     ActionProfiles.createActionProfile(actionProfile, mappingProfileName);
@@ -132,16 +127,16 @@ describe('ui-data-import: Match on Instance identifier match meets both the Iden
     JobProfiles.checkJobProfilePresented(jobProfileName);
 
     cy.visit(TopMenu.dataImportPath);
-    DataImport.uploadFile('marcFileForMatchOnIdentifierForUpdate.mrc', fileNameForUpdateInstance);
+    DataImport.uploadFile(filePathForUpdateInstance, fileNameForUpdateInstance);
     JobProfiles.searchJobProfileForImport(jobProfileName);
     JobProfiles.runImportFile(fileNameForUpdateInstance);
     Logs.checkStatusOfJobProfile('Completed');
     Logs.openFileDetails(fileNameForUpdateInstance);
-    Logs.verifyInstanceStatus(0, 3, 'Updated');
-    Logs.verifyInstanceStatus(1, 3, 'Discarded');
-    Logs.clickOnHotLink(0, 3, 'Updated');
-    InstanceRecordView.verifyMarkAsSuppressedFromDiscovery();
+    Logs.verifyInstanceStatus(0, 3, 'Discarded');
+    Logs.verifyInstanceStatus(1, 3, 'Updated');
+    Logs.clickOnHotLink(1, 3, 'Updated');
     InstanceRecordView.verifyInstanceStatusTerm(mappingProfile.instanceStatusUI);
+    InstanceRecordView.verifyMarkAsSuppressed();
     InstanceRecordView.verifyCatalogedDate(mappingProfile.catalogedDateUI);
     InstanceRecordView.verifyGeneralNoteContent(instanceGeneralNote);
   });
