+++ cypress/e2e/data-import/e2e/match-on-pol-and-update-items.cy.js
@@ -1,38 +1,38 @@
-import getRandomPostfix from '../../support/utils/stringTools';
-import permissions from '../../support/dictionary/permissions';
-import TestTypes from '../../support/dictionary/testTypes';
-import TopMenu from '../../support/fragments/topMenu';
-import NewOrder from '../../support/fragments/orders/newOrder';
-import Orders from '../../support/fragments/orders/orders';
-import Helper from '../../support/fragments/finance/financeHelper';
-import FieldMappingProfiles from '../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
-import NewActionProfile from '../../support/fragments/data_import/action_profiles/newActionProfile';
-import NewFieldMappingProfile from '../../support/fragments/data_import/mapping_profiles/newFieldMappingProfile';
-import ActionProfiles from '../../support/fragments/data_import/action_profiles/actionProfiles';
-import NewJobProfile from '../../support/fragments/data_import/job_profiles/newJobProfile';
-import SettingsMenu from '../../support/fragments/settingsMenu';
-import Users from '../../support/fragments/users/users';
-import MatchProfiles from '../../support/fragments/data_import/match_profiles/matchProfiles';
-import JobProfiles from '../../support/fragments/data_import/job_profiles/jobProfiles';
-import DataImport from '../../support/fragments/data_import/dataImport';
-import Logs from '../../support/fragments/data_import/logs/logs';
-import ItemRecordView from '../../support/fragments/inventory/itemRecordView';
-import CheckInActions from '../../support/fragments/check-in-actions/checkInActions';
-import ServicePoints from '../../support/fragments/settings/tenant/servicePoints/servicePoints';
-import OrderView from '../../support/fragments/orders/orderView';
-import BasicOrderLine from '../../support/fragments/orders/basicOrderLine';
-import Receiving from '../../support/fragments/receiving/receiving';
-import FileDetails from '../../support/fragments/data_import/logs/fileDetails';
-import InventoryInstance from '../../support/fragments/inventory/inventoryInstance';
-import HoldingsRecordView from '../../support/fragments/inventory/holdingsRecordView';
-import ItemView from '../../support/fragments/inventory/inventoryItem/itemView';
-import InventoryViewSource from '../../support/fragments/inventory/inventoryViewSource';
-import NewMatchProfile from '../../support/fragments/data_import/match_profiles/newMatchProfile';
-import Organizations from '../../support/fragments/organizations/organizations';
-import DevTeams from '../../support/dictionary/devTeams';
-import OrderLines from '../../support/fragments/orders/orderLines';
-import NewLocation from '../../support/fragments/settings/tenant/locations/newLocation';
-import FileManager from '../../support/utils/fileManager';
+import getRandomPostfix from '../../../support/utils/stringTools';
+import permissions from '../../../support/dictionary/permissions';
+import TestTypes from '../../../support/dictionary/testTypes';
+import TopMenu from '../../../support/fragments/topMenu';
+import NewOrder from '../../../support/fragments/orders/newOrder';
+import Orders from '../../../support/fragments/orders/orders';
+import Helper from '../../../support/fragments/finance/financeHelper';
+import FieldMappingProfiles from '../../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
+import NewActionProfile from '../../../support/fragments/data_import/action_profiles/newActionProfile';
+import NewFieldMappingProfile from '../../../support/fragments/data_import/mapping_profiles/newFieldMappingProfile';
+import ActionProfiles from '../../../support/fragments/data_import/action_profiles/actionProfiles';
+import NewJobProfile from '../../../support/fragments/data_import/job_profiles/newJobProfile';
+import SettingsMenu from '../../../support/fragments/settingsMenu';
+import Users from '../../../support/fragments/users/users';
+import MatchProfiles from '../../../support/fragments/data_import/match_profiles/matchProfiles';
+import JobProfiles from '../../../support/fragments/data_import/job_profiles/jobProfiles';
+import DataImport from '../../../support/fragments/data_import/dataImport';
+import Logs from '../../../support/fragments/data_import/logs/logs';
+import ItemRecordView from '../../../support/fragments/inventory/itemRecordView';
+import CheckInActions from '../../../support/fragments/check-in-actions/checkInActions';
+import ServicePoints from '../../../support/fragments/settings/tenant/servicePoints/servicePoints';
+import OrderView from '../../../support/fragments/orders/orderView';
+import BasicOrderLine from '../../../support/fragments/orders/basicOrderLine';
+import Receiving from '../../../support/fragments/receiving/receiving';
+import FileDetails from '../../../support/fragments/data_import/logs/fileDetails';
+import InventoryInstance from '../../../support/fragments/inventory/inventoryInstance';
+import HoldingsRecordView from '../../../support/fragments/inventory/holdingsRecordView';
+import ItemView from '../../../support/fragments/inventory/inventoryItem/itemView';
+import InventoryViewSource from '../../../support/fragments/inventory/inventoryViewSource';
+import NewMatchProfile from '../../../support/fragments/data_import/match_profiles/newMatchProfile';
+import Organizations from '../../../support/fragments/organizations/organizations';
+import DevTeams from '../../../support/dictionary/devTeams';
+import OrderLines from '../../../support/fragments/orders/orderLines';
+import NewLocation from '../../../support/fragments/settings/tenant/locations/newLocation';
+import FileManager from '../../../support/utils/fileManager';
 
 describe('ui-data-import: Match on POL and update related Instance, Holdings, Item', () => {
   const firstItem = {
