+++ cypress/e2e/data-import/e2e/match-on-vrn-and-update-items.cy.js
@@ -1,30 +1,30 @@
 /* eslint-disable cypress/no-unnecessary-waiting */
 import uuid from 'uuid';
-import getRandomPostfix from '../../support/utils/stringTools';
-import TestTypes from '../../support/dictionary/testTypes';
-import Helper from '../../support/fragments/finance/financeHelper';
-import TopMenu from '../../support/fragments/topMenu';
-import Logs from '../../support/fragments/data_import/logs/logs';
-import MatchOnVRN from '../../support/fragments/data_import/matchOnVRN';
-import FileManager from '../../support/utils/fileManager';
-import SettingsMenu from '../../support/fragments/settingsMenu';
-import Orders from '../../support/fragments/orders/orders';
-import Users from '../../support/fragments/users/users';
-import permissions from '../../support/dictionary/permissions';
-import DevTeams from '../../support/dictionary/devTeams';
-import JobProfiles from '../../support/fragments/data_import/job_profiles/jobProfiles';
-import BasicOrderLine from '../../support/fragments/orders/basicOrderLine';
-import NewOrder from '../../support/fragments/orders/newOrder';
-import Organizations from '../../support/fragments/organizations/organizations';
-import OrderView from '../../support/fragments/orders/orderView';
-import OrderLines from '../../support/fragments/orders/orderLines';
-import Receiving from '../../support/fragments/receiving/receiving';
-import DataImport from '../../support/fragments/data_import/dataImport';
-import FileDetails from '../../support/fragments/data_import/logs/fileDetails';
-import InventoryInstance from '../../support/fragments/inventory/inventoryInstance';
-import MatchProfiles from '../../support/fragments/data_import/match_profiles/matchProfiles';
-import ActionProfiles from '../../support/fragments/data_import/action_profiles/actionProfiles';
-import FieldMappingProfiles from '../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
+import getRandomPostfix from '../../../support/utils/stringTools';
+import TestTypes from '../../../support/dictionary/testTypes';
+import Helper from '../../../support/fragments/finance/financeHelper';
+import TopMenu from '../../../support/fragments/topMenu';
+import Logs from '../../../support/fragments/data_import/logs/logs';
+import MatchOnVRN from '../../../support/fragments/data_import/matchOnVRN';
+import FileManager from '../../../support/utils/fileManager';
+import SettingsMenu from '../../../support/fragments/settingsMenu';
+import Orders from '../../../support/fragments/orders/orders';
+import Users from '../../../support/fragments/users/users';
+import permissions from '../../../support/dictionary/permissions';
+import DevTeams from '../../../support/dictionary/devTeams';
+import JobProfiles from '../../../support/fragments/data_import/job_profiles/jobProfiles';
+import BasicOrderLine from '../../../support/fragments/orders/basicOrderLine';
+import NewOrder from '../../../support/fragments/orders/newOrder';
+import Organizations from '../../../support/fragments/organizations/organizations';
+import OrderView from '../../../support/fragments/orders/orderView';
+import OrderLines from '../../../support/fragments/orders/orderLines';
+import Receiving from '../../../support/fragments/receiving/receiving';
+import DataImport from '../../../support/fragments/data_import/dataImport';
+import FileDetails from '../../../support/fragments/data_import/logs/fileDetails';
+import InventoryInstance from '../../../support/fragments/inventory/inventoryInstance';
+import MatchProfiles from '../../../support/fragments/data_import/match_profiles/matchProfiles';
+import ActionProfiles from '../../../support/fragments/data_import/action_profiles/actionProfiles';
+import FieldMappingProfiles from '../../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
 
 describe('ui-data-import: Match on VRN and update related Instance, Holdings, Item', () => {
   const item = {
