+++ cypress/e2e/data-import/e2e/mrc-import-file-for-create-items.cy.js
@@ -1,19 +1,19 @@
-import FieldMappingProfiles from '../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
-import NewActionProfile from '../../support/fragments/data_import/action_profiles/newActionProfile';
-import NewFieldMappingProfile from '../../support/fragments/data_import/mapping_profiles/newFieldMappingProfile';
-import ActionProfiles from '../../support/fragments/data_import/action_profiles/actionProfiles';
-import NewJobProfile from '../../support/fragments/data_import/job_profiles/newJobProfile';
-import getRandomPostfix from '../../support/utils/stringTools';
-import DataImport from '../../support/fragments/data_import/dataImport';
-import Logs from '../../support/fragments/data_import/logs/logs';
-import JobProfiles from '../../support/fragments/data_import/job_profiles/jobProfiles';
-import TestTypes from '../../support/dictionary/testTypes';
-import SettingsMenu from '../../support/fragments/settingsMenu';
-import FileDetails from '../../support/fragments/data_import/logs/fileDetails';
-import TopMenu from '../../support/fragments/topMenu';
-import permissions from '../../support/dictionary/permissions';
-import Users from '../../support/fragments/users/users';
-import DevTeams from '../../support/dictionary/devTeams';
+import FieldMappingProfiles from '../../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
+import NewActionProfile from '../../../support/fragments/data_import/action_profiles/newActionProfile';
+import NewFieldMappingProfile from '../../../support/fragments/data_import/mapping_profiles/newFieldMappingProfile';
+import ActionProfiles from '../../../support/fragments/data_import/action_profiles/actionProfiles';
+import NewJobProfile from '../../../support/fragments/data_import/job_profiles/newJobProfile';
+import getRandomPostfix from '../../../support/utils/stringTools';
+import DataImport from '../../../support/fragments/data_import/dataImport';
+import Logs from '../../../support/fragments/data_import/logs/logs';
+import JobProfiles from '../../../support/fragments/data_import/job_profiles/jobProfiles';
+import TestTypes from '../../../support/dictionary/testTypes';
+import SettingsMenu from '../../../support/fragments/settingsMenu';
+import FileDetails from '../../../support/fragments/data_import/logs/fileDetails';
+import TopMenu from '../../../support/fragments/topMenu';
+import permissions from '../../../support/dictionary/permissions';
+import Users from '../../../support/fragments/users/users';
+import DevTeams from '../../../support/dictionary/devTeams';
 
 describe('ui-data-import: MARC file import with creating of the new instance, holding and item', () => {
   let user = {};
