+++ cypress/e2e/data-import/e2e/mrc-import-file-for-modify-marcBib-record.cy.js
@@ -1,23 +1,23 @@
-import TestTypes from '../../support/dictionary/testTypes';
-import FieldMappingProfiles from '../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
-import getRandomPostfix from '../../support/utils/stringTools';
-import NewFieldMappingProfile from '../../support/fragments/data_import/mapping_profiles/newFieldMappingProfile';
-import ActionProfiles from '../../support/fragments/data_import/action_profiles/actionProfiles';
-import NewJobProfile from '../../support/fragments/data_import/job_profiles/newJobProfile';
-import JobProfiles from '../../support/fragments/data_import/job_profiles/jobProfiles';
-import MatchProfiles from '../../support/fragments/data_import/match_profiles/matchProfiles';
-import DataImport from '../../support/fragments/data_import/dataImport';
-import Logs from '../../support/fragments/data_import/logs/logs';
-import InventorySearchAndFilter from '../../support/fragments/inventory/inventorySearchAndFilter';
-import ExportMarcFile from '../../support/fragments/data-export/export-marc-file';
-import FileManager from '../../support/utils/fileManager';
-import ExportFile from '../../support/fragments/data-export/exportFile';
-import SettingsMenu from '../../support/fragments/settingsMenu';
-import FileDetails from '../../support/fragments/data_import/logs/fileDetails';
-import TopMenu from '../../support/fragments/topMenu';
-import permissions from '../../support/dictionary/permissions';
-import Users from '../../support/fragments/users/users';
-import DevTeams from '../../support/dictionary/devTeams';
+import TestTypes from '../../../support/dictionary/testTypes';
+import FieldMappingProfiles from '../../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
+import getRandomPostfix from '../../../support/utils/stringTools';
+import NewFieldMappingProfile from '../../../support/fragments/data_import/mapping_profiles/newFieldMappingProfile';
+import ActionProfiles from '../../../support/fragments/data_import/action_profiles/actionProfiles';
+import NewJobProfile from '../../../support/fragments/data_import/job_profiles/newJobProfile';
+import JobProfiles from '../../../support/fragments/data_import/job_profiles/jobProfiles';
+import MatchProfiles from '../../../support/fragments/data_import/match_profiles/matchProfiles';
+import DataImport from '../../../support/fragments/data_import/dataImport';
+import Logs from '../../../support/fragments/data_import/logs/logs';
+import InventorySearchAndFilter from '../../../support/fragments/inventory/inventorySearchAndFilter';
+import ExportMarcFile from '../../../support/fragments/data-export/export-marc-file';
+import FileManager from '../../../support/utils/fileManager';
+import ExportFile from '../../../support/fragments/data-export/exportFile';
+import SettingsMenu from '../../../support/fragments/settingsMenu';
+import FileDetails from '../../../support/fragments/data_import/logs/fileDetails';
+import TopMenu from '../../../support/fragments/topMenu';
+import permissions from '../../../support/dictionary/permissions';
+import Users from '../../../support/fragments/users/users';
+import DevTeams from '../../../support/dictionary/devTeams';
 
 describe('ui-data-import: Verify the possibility to modify MARC Bibliographic record', () => {
   // unique name for profiles
