+++ cypress/e2e/data-import/e2e/mrc-import-file-for-update-items.cy.js
@@ -1,26 +1,26 @@
-import TestTypes from '../../support/dictionary/testTypes';
-import getRandomPostfix from '../../support/utils/stringTools';
-import DataImport from '../../support/fragments/data_import/dataImport';
-import JobProfiles from '../../support/fragments/data_import/job_profiles/jobProfiles';
-import Logs from '../../support/fragments/data_import/logs/logs';
-import InventorySearchAndFilter from '../../support/fragments/inventory/inventorySearchAndFilter';
-import ExportFile from '../../support/fragments/data-export/exportFile';
-import TopMenu from '../../support/fragments/topMenu';
-import ExportMarcFile from '../../support/fragments/data-export/export-marc-file';
-import MatchProfiles from '../../support/fragments/data_import/match_profiles/matchProfiles';
-import NewMatchProfile from '../../support/fragments/data_import/match_profiles/newMatchProfile';
-import NewFieldMappingProfile from '../../support/fragments/data_import/mapping_profiles/newFieldMappingProfile';
-import NewActionProfile from '../../support/fragments/data_import/action_profiles/newActionProfile';
-import FieldMappingProfiles from '../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
-import ActionProfiles from '../../support/fragments/data_import/action_profiles/actionProfiles';
-import NewJobProfile from '../../support/fragments/data_import/job_profiles/newJobProfile';
-import FileManager from '../../support/utils/fileManager';
-import ExportFieldMappingProfiles from '../../support/fragments/data-export/exportMappingProfile/exportFieldMappingProfiles';
-import ExportJobProfiles from '../../support/fragments/data-export/exportJobProfile/exportJobProfiles';
-import SettingsMenu from '../../support/fragments/settingsMenu';
-import FileDetails from '../../support/fragments/data_import/logs/fileDetails';
-import SettingsJobProfiles from '../../support/fragments/settings/dataImport/settingsJobProfiles';
-import DevTeams from '../../support/dictionary/devTeams';
+import TestTypes from '../../../support/dictionary/testTypes';
+import getRandomPostfix from '../../../support/utils/stringTools';
+import DataImport from '../../../support/fragments/data_import/dataImport';
+import JobProfiles from '../../../support/fragments/data_import/job_profiles/jobProfiles';
+import Logs from '../../../support/fragments/data_import/logs/logs';
+import InventorySearchAndFilter from '../../../support/fragments/inventory/inventorySearchAndFilter';
+import ExportFile from '../../../support/fragments/data-export/exportFile';
+import TopMenu from '../../../support/fragments/topMenu';
+import ExportMarcFile from '../../../support/fragments/data-export/export-marc-file';
+import MatchProfiles from '../../../support/fragments/data_import/match_profiles/matchProfiles';
+import NewMatchProfile from '../../../support/fragments/data_import/match_profiles/newMatchProfile';
+import NewFieldMappingProfile from '../../../support/fragments/data_import/mapping_profiles/newFieldMappingProfile';
+import NewActionProfile from '../../../support/fragments/data_import/action_profiles/newActionProfile';
+import FieldMappingProfiles from '../../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
+import ActionProfiles from '../../../support/fragments/data_import/action_profiles/actionProfiles';
+import NewJobProfile from '../../../support/fragments/data_import/job_profiles/newJobProfile';
+import FileManager from '../../../support/utils/fileManager';
+import ExportFieldMappingProfiles from '../../../support/fragments/data-export/exportMappingProfile/exportFieldMappingProfiles';
+import ExportJobProfiles from '../../../support/fragments/data-export/exportJobProfile/exportJobProfiles';
+import SettingsMenu from '../../../support/fragments/settingsMenu';
+import FileDetails from '../../../support/fragments/data_import/logs/fileDetails';
+import SettingsJobProfiles from '../../../support/fragments/settings/dataImport/settingsJobProfiles';
+import DevTeams from '../../../support/dictionary/devTeams';
 
 describe('ui-data-import: MARC file upload with the update of instance, holding, and items', () => {
   // profile names for creating
