+++ cypress/e2e/data-import/e2e/mrc-import-with-matching-001.cy.js
@@ -1,24 +1,24 @@
-import DataImport from '../../support/fragments/data_import/dataImport';
-import JobProfiles from '../../support/fragments/data_import/job_profiles/jobProfiles';
-import TopMenu from '../../support/fragments/topMenu';
-import MatchProfiles from '../../support/fragments/data_import/match_profiles/matchProfiles';
-import Logs from '../../support/fragments/data_import/logs/logs';
-import InventorySearchAndFilter from '../../support/fragments/inventory/inventorySearchAndFilter';
-import getRandomPostfix from '../../support/utils/stringTools';
-import NewFieldMappingProfile from '../../support/fragments/data_import/mapping_profiles/newFieldMappingProfile';
-import FieldMappingProfiles from '../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
-import NewActionProfile from '../../support/fragments/data_import/action_profiles/newActionProfile';
-import ActionProfiles from '../../support/fragments/data_import/action_profiles/actionProfiles';
-import NewJobProfile from '../../support/fragments/data_import/job_profiles/newJobProfile';
-import ExportFile from '../../support/fragments/data-export/exportFile';
-import ExportMarcFile from '../../support/fragments/data-export/export-marc-file';
-import FileManager from '../../support/utils/fileManager';
-import TestTypes from '../../support/dictionary/testTypes';
-import SettingsMenu from '../../support/fragments/settingsMenu';
-import FileDetails from '../../support/fragments/data_import/logs/fileDetails';
-import permissions from '../../support/dictionary/permissions';
-import Users from '../../support/fragments/users/users';
-import DevTeams from '../../support/dictionary/devTeams';
+import DataImport from '../../../support/fragments/data_import/dataImport';
+import JobProfiles from '../../../support/fragments/data_import/job_profiles/jobProfiles';
+import TopMenu from '../../../support/fragments/topMenu';
+import MatchProfiles from '../../../support/fragments/data_import/match_profiles/matchProfiles';
+import Logs from '../../../support/fragments/data_import/logs/logs';
+import InventorySearchAndFilter from '../../../support/fragments/inventory/inventorySearchAndFilter';
+import getRandomPostfix from '../../../support/utils/stringTools';
+import NewFieldMappingProfile from '../../../support/fragments/data_import/mapping_profiles/newFieldMappingProfile';
+import FieldMappingProfiles from '../../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
+import NewActionProfile from '../../../support/fragments/data_import/action_profiles/newActionProfile';
+import ActionProfiles from '../../../support/fragments/data_import/action_profiles/actionProfiles';
+import NewJobProfile from '../../../support/fragments/data_import/job_profiles/newJobProfile';
+import ExportFile from '../../../support/fragments/data-export/exportFile';
+import ExportMarcFile from '../../../support/fragments/data-export/export-marc-file';
+import FileManager from '../../../support/utils/fileManager';
+import TestTypes from '../../../support/dictionary/testTypes';
+import SettingsMenu from '../../../support/fragments/settingsMenu';
+import FileDetails from '../../../support/fragments/data_import/logs/fileDetails';
+import permissions from '../../../support/dictionary/permissions';
+import Users from '../../../support/fragments/users/users';
+import DevTeams from '../../../support/dictionary/devTeams';
 
 describe('ui-data-import: MARC-MARC matching for 001 field', () => {
   let user = {};
