+++ cypress/e2e/data-import/e2e/mrc-import-with-matching-999ff.cy.js
@@ -1,22 +1,22 @@
-import FieldMappingProfiles from '../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
-import NewActionProfile from '../../support/fragments/data_import/action_profiles/newActionProfile';
-import NewFieldMappingProfile from '../../support/fragments/data_import/mapping_profiles/newFieldMappingProfile';
-import ActionProfiles from '../../support/fragments/data_import/action_profiles/actionProfiles';
-import NewJobProfile from '../../support/fragments/data_import/job_profiles/newJobProfile';
-import MatchProfiles from '../../support/fragments/data_import/match_profiles/matchProfiles';
-import DataImport from '../../support/fragments/data_import/dataImport';
-import Logs from '../../support/fragments/data_import/logs/logs';
-import JobProfiles from '../../support/fragments/data_import/job_profiles/jobProfiles';
-import TestTypes from '../../support/dictionary/testTypes';
-import InventorySearchAndFilter from '../../support/fragments/inventory/inventorySearchAndFilter';
-import ExportFile from '../../support/fragments/data-export/exportFile';
-import ExportMarcFile from '../../support/fragments/data-export/export-marc-file';
-import FileManager from '../../support/utils/fileManager';
-import getRandomPostfix from '../../support/utils/stringTools';
-import SettingsMenu from '../../support/fragments/settingsMenu';
-import FileDetails from '../../support/fragments/data_import/logs/fileDetails';
-import TopMenu from '../../support/fragments/topMenu';
-import DevTeams from '../../support/dictionary/devTeams';
+import FieldMappingProfiles from '../../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
+import NewActionProfile from '../../../support/fragments/data_import/action_profiles/newActionProfile';
+import NewFieldMappingProfile from '../../../support/fragments/data_import/mapping_profiles/newFieldMappingProfile';
+import ActionProfiles from '../../../support/fragments/data_import/action_profiles/actionProfiles';
+import NewJobProfile from '../../../support/fragments/data_import/job_profiles/newJobProfile';
+import MatchProfiles from '../../../support/fragments/data_import/match_profiles/matchProfiles';
+import DataImport from '../../../support/fragments/data_import/dataImport';
+import Logs from '../../../support/fragments/data_import/logs/logs';
+import JobProfiles from '../../../support/fragments/data_import/job_profiles/jobProfiles';
+import TestTypes from '../../../support/dictionary/testTypes';
+import InventorySearchAndFilter from '../../../support/fragments/inventory/inventorySearchAndFilter';
+import ExportFile from '../../../support/fragments/data-export/exportFile';
+import ExportMarcFile from '../../../support/fragments/data-export/export-marc-file';
+import FileManager from '../../../support/utils/fileManager';
+import getRandomPostfix from '../../../support/utils/stringTools';
+import SettingsMenu from '../../../support/fragments/settingsMenu';
+import FileDetails from '../../../support/fragments/data_import/logs/fileDetails';
+import TopMenu from '../../../support/fragments/topMenu';
+import DevTeams from '../../../support/dictionary/devTeams';
 
 describe('ui-data-import: MARC file import with matching for 999 ff field', () => {
   // unique file name to upload
