+++ cypress/e2e/data-import/e2e/protected-fields-in-incoming-records-not-deleted-scen1.cy.js
@@ -1,20 +1,20 @@
-import getRandomPostfix from '../../support/utils/stringTools';
-import permissions from '../../support/dictionary/permissions';
-import TestTypes from '../../support/dictionary/testTypes';
-import DevTeams from '../../support/dictionary/devTeams';
-import TopMenu from '../../support/fragments/topMenu';
-import SettingsMenu from '../../support/fragments/settingsMenu';
-import DataImport from '../../support/fragments/data_import/dataImport';
-import JobProfiles from '../../support/fragments/data_import/job_profiles/jobProfiles';
-import Logs from '../../support/fragments/data_import/logs/logs';
-import FileDetails from '../../support/fragments/data_import/logs/fileDetails';
-import InventorySearchAndFilter from '../../support/fragments/inventory/inventorySearchAndFilter';
-import MarcFieldProtection from '../../support/fragments/settings/dataImport/marcFieldProtection';
-import Z3950TargetProfiles from '../../support/fragments/settings/inventory/z39.50TargetProfiles';
-import InventoryInstance from '../../support/fragments/inventory/inventoryInstance';
-import InventoryEditMarcRecord from '../../support/fragments/inventory/inventoryEditMarcRecord';
-import InventoryViewSource from '../../support/fragments/inventory/inventoryViewSource';
-import Users from '../../support/fragments/users/users';
+import getRandomPostfix from '../../../support/utils/stringTools';
+import permissions from '../../../support/dictionary/permissions';
+import TestTypes from '../../../support/dictionary/testTypes';
+import DevTeams from '../../../support/dictionary/devTeams';
+import TopMenu from '../../../support/fragments/topMenu';
+import SettingsMenu from '../../../support/fragments/settingsMenu';
+import DataImport from '../../../support/fragments/data_import/dataImport';
+import JobProfiles from '../../../support/fragments/data_import/job_profiles/jobProfiles';
+import Logs from '../../../support/fragments/data_import/logs/logs';
+import FileDetails from '../../../support/fragments/data_import/logs/fileDetails';
+import InventorySearchAndFilter from '../../../support/fragments/inventory/inventorySearchAndFilter';
+import MarcFieldProtection from '../../../support/fragments/settings/dataImport/marcFieldProtection';
+import Z3950TargetProfiles from '../../../support/fragments/settings/inventory/z39.50TargetProfiles';
+import InventoryInstance from '../../../support/fragments/inventory/inventoryInstance';
+import InventoryEditMarcRecord from '../../../support/fragments/inventory/inventoryEditMarcRecord';
+import InventoryViewSource from '../../../support/fragments/inventory/inventoryViewSource';
+import Users from '../../../support/fragments/users/users';
 
 describe('ui-data-import: Check that protected fields in incoming records are not deleted during import: Scenario 1', () => {
   let user = null;
