+++ cypress/e2e/data-import/importing-marc-bib-files/001-003-005-handling-for-new-and-update-srs-record.cy.js
@@ -1,25 +1,25 @@
-import getRandomPostfix from '../../support/utils/stringTools';
-import TestTypes from '../../support/dictionary/testTypes';
-import DevTeams from '../../support/dictionary/devTeams';
-import TopMenu from '../../support/fragments/topMenu';
-import DataImport from '../../support/fragments/data_import/dataImport';
-import JobProfiles from '../../support/fragments/data_import/job_profiles/jobProfiles';
-import Logs from '../../support/fragments/data_import/logs/logs';
-import FileDetails from '../../support/fragments/data_import/logs/fileDetails';
-import InventoryInstance from '../../support/fragments/inventory/inventoryInstance';
-import NewMatchProfile from '../../support/fragments/data_import/match_profiles/newMatchProfile';
-import Helper from '../../support/fragments/finance/financeHelper';
-import NewFieldMappingProfile from '../../support/fragments/data_import/mapping_profiles/newFieldMappingProfile';
-import SettingsMenu from '../../support/fragments/settingsMenu';
-import MatchProfiles from '../../support/fragments/data_import/match_profiles/matchProfiles';
-import FieldMappingProfiles from '../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
-import NewActionProfile from '../../support/fragments/data_import/action_profiles/newActionProfile';
-import ActionProfiles from '../../support/fragments/data_import/action_profiles/actionProfiles';
-import NewJobProfile from '../../support/fragments/data_import/job_profiles/newJobProfile';
-import InventoryViewSource from '../../support/fragments/inventory/inventoryViewSource';
-import ExportMarcFile from '../../support/fragments/data-export/export-marc-file';
-import InventorySearchAndFilter from '../../support/fragments/inventory/inventorySearchAndFilter';
-import FileManager from '../../support/utils/fileManager';
+import getRandomPostfix from '../../../support/utils/stringTools';
+import TestTypes from '../../../support/dictionary/testTypes';
+import DevTeams from '../../../support/dictionary/devTeams';
+import TopMenu from '../../../support/fragments/topMenu';
+import DataImport from '../../../support/fragments/data_import/dataImport';
+import JobProfiles from '../../../support/fragments/data_import/job_profiles/jobProfiles';
+import Logs from '../../../support/fragments/data_import/logs/logs';
+import FileDetails from '../../../support/fragments/data_import/logs/fileDetails';
+import InventoryInstance from '../../../support/fragments/inventory/inventoryInstance';
+import NewMatchProfile from '../../../support/fragments/data_import/match_profiles/newMatchProfile';
+import Helper from '../../../support/fragments/finance/financeHelper';
+import NewFieldMappingProfile from '../../../support/fragments/data_import/mapping_profiles/newFieldMappingProfile';
+import SettingsMenu from '../../../support/fragments/settingsMenu';
+import MatchProfiles from '../../../support/fragments/data_import/match_profiles/matchProfiles';
+import FieldMappingProfiles from '../../../support/fragments/data_import/mapping_profiles/fieldMappingProfiles';
+import NewActionProfile from '../../../support/fragments/data_import/action_profiles/newActionProfile';
+import ActionProfiles from '../../../support/fragments/data_import/action_profiles/actionProfiles';
+import NewJobProfile from '../../../support/fragments/data_import/job_profiles/newJobProfile';
+import InventoryViewSource from '../../../support/fragments/inventory/inventoryViewSource';
+import ExportMarcFile from '../../../support/fragments/data-export/export-marc-file';
+import InventorySearchAndFilter from '../../../support/fragments/inventory/inventorySearchAndFilter';
+import FileManager from '../../../support/utils/fileManager';
 
 describe('ui-data-import: Test 001/003/035 handling for New and Updated SRS records', () => {
   let instanceHrid = null;
