+++ cypress/support/fragments/data_import/mapping_profiles/fieldMappingProfiles.js
@@ -9,6 +9,7 @@ import {
   Form
 } from '../../../../../interactors';
 import { getLongDelay } from '../../../utils/cypressTools';
+import FieldMappingProfileEdit from './fieldMappingProfileEdit';
 import FieldMappingProfileView from './fieldMappingProfileView';
 import NewFieldMappingProfile from './newFieldMappingProfile';
 
@@ -82,6 +83,7 @@ export default {
   openNewMappingProfileForm,
   saveProfile,
   closeViewModeForMappingProfile,
+  searchMappingProfile,
 
   createMappingProfile:(mappingProfile) => {
     openNewMappingProfileForm();
@@ -156,8 +158,8 @@ export default {
   createMappingProfileForUpdatesAndOverrideMarc:(mappingProfile, firstProtectedField, secondProtectedField) => {
     openNewMappingProfileForm();
     NewFieldMappingProfile.fillMappingProfileForUpdatesMarc(mappingProfile);
-    FieldMappingProfileView.markFieldForProtection(firstProtectedField);
-    FieldMappingProfileView.markFieldForProtection(secondProtectedField);
+    FieldMappingProfileEdit.markFieldForProtection(firstProtectedField);
+    FieldMappingProfileEdit.markFieldForProtection(secondProtectedField);
     cy.do(saveProfileButton.click());
   },
